Extract a factory for helping in tests
